Allow for users to specify Golang and Docker versions and RPM repos
In order to allow each job building an AMI to be able to declare the
Docker and Golang versions that should be present, the Docker and
Golang install paths are separated out so that they live alone and
are not called in the `build-origin-base` command but instead in
specific `install-docker` and `install-golang` commands that can be
run by a user to configfure the version of Docker or Golang installed
on the system.

These new commands have the following flags:
  --{golang,docker}.version to specify the version to install
The following flags determine what repos are enabled during the
installation process:
  --repourl to specify the URL of an RPM repo
  --repo to specify the name of an installed repo

Any repositories installed from URLs  provided with the above flags
will be enabled *only* for the installation of Golang or Docker,
and will be disabled thereafter.

The `--{golang,docker}.version` flags are used to construct package
names for the core package as well as for peripheral packages, for
instance `golang` and `golang-src`. No leading dash is necessary
to specify the version. For instance, a user would specify that they
want `golang-1.6.2` by passing `--golang.version=1.6.2`.

Repositories installed from URLs will have a .repo file created for
them automatically. The name of the repository will be the given
URL with all non-alphanumeric characters (/[^a-zA-Z0-9]/) stripped
out. Then, the URL and name will be inserted into the following
definition:
```
[#{reponame}]
name=#{reponame}
baseurl=#{url}
enabled=1
gpgcheck=0
sslverify=0
sslclientcert=/var/lib/yum/client-cert.pem
sslclientkey=/var/lib/yum/client-key.pem
```

As the sslclient{cert,key} are not used unless the RPM repository
responds to a request with auth headers, these fields are ignored
unless the repository needs them and therefore repositories that
do not need them are not broken and it is safe to append them always.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>