Configure `build-origin-base` to use local scripts
The code that supports `build-origin-base` was becoming unwieldy, so to
allow for easy modification and maintenance, it will be refactored to run
local scripts on the machine. The scripts are packaged with the plugin.

The first step of this command now is to copy all the available scripts
to the machine. The workflow for running scripts is shown in the refactor
of the network configuration script in this commit.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>