Muted SSH exit status errors on test commands
In order to mangle the output of test scripts less,
all test calls to machine.communicator.{sudo,execute}
now pass :error_key => :ssh_bad_exit_status_muted, so
the errors that are printed by Vagrant do not contain
the command that was run, stderr, or stdout. The test
commands we are running will report this information
themselves by using the Origin stacktrace functions.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>