add `--prune=all` to `git gc`
`git gc` leaves 2 weeks of refs by default, but with a heavily-used repo
that means loads of "`dangling commit`" messages in the git fsck output.
Adding `--prune=all` cleans up all commits, so we should only get one or
a couple of dangling commits in the output on the next run.