Remove shell-word escaping from extended test execution code
A valid extended test bucket name should never contain any
special shell characters, so it is unnecessary to attempt
to escape them. Similarly, the only characters inside of a
focus string that we are vulnerable to are double-quotes.
Extended test focuses should not contain literal double
quotes, so we should not need to escape anything in those
either.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>