Customer Diagnostics commands
The customer diagnostics team is prototyping a test suite written in go for
testing ansible-based tooling.  Right now we're still in the early stages so a
fork of openshift-ansible is being used.  If things go well by the end of the
sprint we expect to merge this to openshift-ansible and at that time we'd
update the urls accordingly.

The goal with these new commands is simply to enable a test and merge workflow
that matches other Origin teams.  The implementation is almost entirely similar
to the other sync and test commands that we have.