Added an option to skip Docker volume group setup
In the generic case where a user is creating an environment using this
tool, they may or may not want to use volume groups for Docker storage.
We should allow them to choose, but keep the current behavior as the
default. When we are trying to use volume groups but cannot find the
group to use, we should also fail loudly and clearly.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>