Reconfigure Docker thin-pool setup to use `docker-storage-setup`
Instead of configuring the logical volumes by hand, we should now use
the `docker-storage-setup` tool instead. The manual process is no longer
honored by Docker 1.12.6+. The ratio of data to metadata in the logical
volume is not identical to the ratio previously, but no compelling
reasons are known for which the default from `docker-storage-setup`
would be inappropriate.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>