Make the repo name to URL logic simpler and more generic
It is the case for all of the repositories that we currently support in
this project that the 'repository_name' parameter is literally the name
of the project underneath the OpenShift organization on GitHub, so it is
not necessary to have all of the complicated logic around the repository
name to repository URL mapping that we have today. These changes
simplify that logic and thereby also make it more generic, allowing new
repositories to be supported implicitly for commands like
`checkout-repos` and `origin-local-checkout`.

This is a backwards-incompatible change in that the definition of the
`origin` repository no longer contains source-to-image or the Origin web
console, but it does not seem to actually be necessary to clone these
three repositories in tandem. When a Jenkins job in fact needs more than
one repository cloned or updated, it can call out to do this with the
pre-existing `--repo` flag.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>