Use `origin-web-console` instead of `origin-console`
The Git repository for the Origin Web Console is named
`origin-web-console`, not `origin-console`, so we should use the correct
name to identify it in our code. This is now necessary as we have
refacoted the code to remove the need for nick-names for the repos and
instead use the names of the repos themselves.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>