DRY out repo sync logic, add Jenkins sync
Removing a lot of the duplication between the different repo sync
commands allows us to add one for the Jenkins repo with minimal effort.
It would be nice to have a `vagrant sync-repo [REPO]` command but this
would be a breaking change for the CLI, so we can continue with the
named commands per repo instead.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>